backend/requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.36.2
anyio==4.6.2.post1
assemblyai==0.34.0
asyncio==3.4.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
deepgram-sdk==3.7.4
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
fastapi==0.115.2
filelock==3.16.1
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.0
idna==3.10
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==0.2.3
langchain-core==0.3.12
langchain-openai==0.2.3
langsmith==0.1.136
marshmallow==3.23.0
multidict==6.1.0
mypy-extensions==1.0.0
openai==1.52.0
orjson==3.10.8
packaging==24.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==0.40.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.20.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==13.1
yarl==1.15.5
